requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.8.3
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
bandit==1.7.10
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==25.11.0
boolean.py==5.0
BotCore==1.1.4
boto3==1.41.4
botocore==1.41.4
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.12.0
cryptography==46.0.3
cycler==0.12.1
cyclonedx-python-lib==7.6.2
dacite==1.9.2
datasets==4.4.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
distlib==0.4.0
dm-tree==0.1.9
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.1
etils==1.13.0
fiftyone==1.10.0
fiftyone-brain==0.21.4
fiftyone_db==1.4.0
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gast==0.6.0
glob2==0.7
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hf-xet==1.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
humanize==4.14.0
Hypercorn==0.18.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
ImageIO==2.37.2
immutabledict==4.2.2
importlib_resources==6.5.2
inflate64==1.0.3
iniconfig==2.3.0
interrogate==1.7.0
isort==7.0.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
kiwisolver==1.4.9
lazy_loader==0.4
lia-web==0.2.3
libclang==18.1.1
license-expression==30.4.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.19
ml_dtypes==0.5.4
mongoengine==0.29.1
motor==3.6.1
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
multivolumefile==0.2.3
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.12.0
networkx==3.6
nodeenv==1.9.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packageurl-python==0.17.6
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
pathspec==0.12.1
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
pprintpp==0.4.0
pre_commit==4.5.0
priority==2.0.0
promise==2.3
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
py==1.11.0
py-serializable==1.1.2
py7zr==1.0.0
pyarrow==22.0.0
pybcj==1.0.6
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydash==8.0.5
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.6.1
pyparsing==3.2.5
pyppmd==1.2.0
pyright==1.1.389
pyTelegramBotAPI==4.29.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzstd==0.18.0
rarfile==4.2
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rtree==1.4.1
ruff==0.6.9
s3transfer==0.15.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.50.0
stevedore==5.6.0
strawberry-graphql==0.287.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
termcolor==3.2.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.10.16
toml==0.10.2
tqdm==4.67.1
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.5.0
virtualenv==20.35.4
voxel51-eta==0.15.1
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.2
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
